requirements.txt
python-telegram-bot[job-queue]==21.10
anthropic>=0.42.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
matplotlib>=3.9.0
